Use the MultiThreadedDebugDLL runtime in the Debug|x64 confíguraion
The core build must then have been configured with
--enable-msvc-debug-runtime.

Yes, it is silly that these have to be kept manually in sync. I should
change the code here in COOL to call lok::Office::freeError() (for
which I today added a wrapper with the better name freeMemory()) on
dynamic memory allocated by LOKit, then it doesn't matter.

Signed-off-by: Tor Lillqvist <tml@collabora.com>
Change-Id: I14190f6a66313a88345225561b0cc365bfc8955e