Add a simple command-line tool to manipulate the Windows clipboard
To be used in experimentation, for instance to figure out what formats
actually are relevant and used in different scenarios by various
software. In particular, it can be useful to be able to remove
specific data formats from the clipboard.

When removing formats, note that Windows automatically synthesizes
some formats even if they haven't been explicitly stored on the
clipboard. See
https://learn.microsoft.com/en-us/windows/win32/dataxchg/clipboard-formats#synthesized-clipboard-formats
If you try to use this tool to remove CF_TEXT data, the CF_UNICODETEXT
data will remain and CF_TEXT will be synthesized from that and be
still present.

Usage:
clipboard help                  -- show this text
clipboard list                  -- list formats on the clipboard
clipboard dump fmt              -- output data for a format to stdout
clipboard show fmt              -- show data for a format as hex dump
clipboard drop fmt1,fmt2,...    -- drop a set of formats from the clipboard
clipboard keep fmt1,fmt2,...    -- keep only a set of formats on the clipboard

Signed-off-by: Tor Lillqvist <tml@collabora.com>
Change-Id: Ia18804f79154546b5521915491946d44bdb99176