Add a few Windows clipboard formats that already are MIME types
Also, make sure not to pass the same MIME type multiple times to
lok::Document::setClipboard(). Sure, in theory it is possible that
some application stores *different* data for instance for the "PNG"
and "image/png" clipboard formats, but shame on it then. We are not
going to guess which one is better.

Signed-off-by: Tor Lillqvist <tml@collabora.com>
Change-Id: Idae2974cac8827fe005ad40065e394bf84964693