Add a separate tiny program to print a PDF file to CODA-W
It is fairly trivial to print a PDF from managed C# code. But sadly
not, as far as I know, from unmanaged Win32 C++ code, which is what
CODA-W now is. So just use a tiny separate C# program to print.

This commit only adds the program (as a separate project in the CODA
solution). I will add code to CODA.cpp to run it next.

Signed-off-by: Tor Lillqvist <tml@collabora.com>
Change-Id: I7aeafd472516c21556f50cb7eb5ccdd6e3c1e13b