Fix and simplify logging in CODA-W
Just log the basename of the source file on Windows, too.

Drop some pointless leftover test LOG_TRC() calls

Split out the Log API implementation for Windows into a separate
Log-windows.cpp file. Note that Poco is not used in it.

Logging is a no-op in CODA-W unless the process is being debugged. Log
output is via OutputDebugString(). (Standard outoput and standard
error doesn't show up anywhere for a GUI program on Windows, unless
expicitly started from the command line with redirection to a file or
pipe.)

Keep the existing Poco-based code for other platforms, but rename
Log.cpp to Log-poco.cpp. Add a small Log-common.cpp for the
Log::prefix() function.

Now all logging shows up in CODA-W. The Poco logging setup from
Poco::Application was screwing things up thanks to COOLWSD being
derived from that class, which insists on doing its own things, which
overrode what our code had done, and caused the Poco::ConsoleChannel
to be used instead of our Log::ConsoleChannel. Apparently in
server-based COOL the call to Log::initialize() helps, but that call
is inside a !MOBILEAPP ifdef, and we have already called
Log::initialize() right at the start in CODA.cpp anyway.

I must admit our Log code is so complicated with all the Logger and
Channel things that I couldn't understand it fully. But anyway,
getting rid of those Poco things for logging on Windows was a relief,
that code is minimal and clean.

(I didn't adapt the iOS, Android, and macOS projects to the renamed
and additional source files, sorry. But should be trivial. I did check
that this didn't break anything on Linux.)

Signed-off-by: Tor Lillqvist <tml@collabora.com>
Change-Id: I8b6d741dcb3f6713a8f3879cf12b38dcfe52c89f