CODA-Q: Use core's default UserInstallation
Currently, at least a flatpak build would fail to start with

> Collabora Office 25.04 - Fatal Error: The application cannot be started.
> User installation could not be completed.

caused by this code causing lo_initialize (in core desktop/source/lib/init.cxx)
to be called with a non-null pUserProfileUrl argument, which thus calls

> rtl::Bootstrap::set(u"UserInstallation"_ustr, url);

so that the LO code then uses a UserInstallation value of
"file:///app/core/instdir", thus tries to mkdir an /app/core/instdir/user,
failing with EROFS, causing the above fatal error message.

But using a null pUserProfileUrl should be OK here:  The code in code
lo_initialize (desktop/source/lib/init.cxx) will not explicitly
rtl::Bootstrap::set UserInstallation, so it will use the default from the LO
installation's program/bootstraprc, i.e.,
UserInstallation=$SYSUSERCONFIG/collaboraoffice/4, i.e., at the
$XDG_CONFIG_HOME/collaboraoffice/4 pathname.

For the flatpak app, $XDG_CONFIG_HOME/collaboraoffice/4 resolves to something
like the per-flatpak-app pathname
~/.var/app/org.collabora.CODA/config/collaboraoffice/4 in the (host) filesystem,
so could not clash with a user installation of any other (LO-derived)
application that also happens to be installed on the host machine.

And for the non-flatpak case, $XDG_CONFIG_HOME/collaboraoffice/4 resolves to
something like ~/.config/collaboraoffice/4, which should hopefully also not
clash with a user installation of any other (LO-derived) application that also
happens to be installed on the machine.

(However, there is of course the caveat with multiple instances of either the
flatpak app or the non-flatpak application running at the same time, which I
described elsewhere:  "So yeah, UserInstallation appears to be a general problem
with all of the CODA apps:  The LO code is not designed in a way that multiple
instances of it can concurrently share a user installation, that's why e.g. LO's
soffice early on checks whether there's already another soffice instance running
for the same UserInstallation, and passes its command line args to that other
instance and terminates if that's the case.  For CODA apps, there's apparently
no such countermeasures, so one instance of such an app could mess with another
instance or even with a LO installation (when the CODA app would call lok_init_2
with a null user_profile_url argument, so would default to the UserInstallation
value from LO's bootstrap ini-file, IIUC)")

Signed-off-by: Stephan Bergmann <stephan.bergmann@collabora.com>
Change-Id: I81a3f03fed1a82b1b9548286b93b90920def26ed