Clean up document and app lifecycle in CODA-W
Reduce the amount of ad-hoc code. Let the code in kit call
destroyView(), do not do that in CODA.cpp. Do not call
TerminateProcess(), but let innerMain() call Util::forcedExit() which
calls _Exit() instead. Which also is a somewhat blunt exit, but anyway
it is cross-platorm for MOBILEAPP.

Also, close the persistentWindowSizeStore when about to exit.

Signed-off-by: Tor Lillqvist <tml@collabora.com>
Change-Id: I252849a5c0f9a12965e6ebaa55d6a6748b506811