FakeSocket - detect and warn on racing connections.
In rare cases we can get adjacent back to back connections:

09:41:58.728480 INF  Connecting to 8| net/Socket.cpp
09:41:58.728482 INF  Connecting to 8| net/Socket.cpp
09:41:58.728506 ERR  Failed to connect to the 'wsd' socket| net/Socket.cpp

Where the listen/accept has not managed to process the incoming
connect before the next connect arrives and fails.

So instead in this (hopefully rare case) - block on the accept,
not ideal - but simple at least afterwards:

09:41:58.728480 INF  Connecting to 8| net/Socket.cpp
09:41:58.728482 INF  Connecting to 8| net/Socket.cpp
22:05:57.681707 [ prisoner_poll ] TRC  #8: Accepted client #29, StreamSocket[#29, HTTP, Unix @ :0]| net/ServerSocket.hpp
22:05:57.681801 [ prisoner_poll ] TRC  #8: Accepted client #31, StreamSocket[#31, HTTP, Unix @ :0]| net/ServerSocket.hpp

Which seems more helpful.

Signed-off-by: Michael Meeks <michael.meeks@collabora.com>
Change-Id: I6b1295f2c73b5c012068161adf04330821c9294e