Introduce an "isnewdocument" parameter to cool.html
Add window.mode.isNewDocument() to check for it.

Use that to not turn on the read-only mode in CODA when it is in fact
a completely new document that is being edited. It does not make sense
to open a such in read-only mode.

Pass that parameter in CODA-W, for the new document buttons in the
file open dialog.

Signed-off-by: Tor Lillqvist <tml@collabora.com>
Change-Id: Ie171fe4142d14bdd98f9966f98bdf4f6ebd1d133