CODA-Q: prevent unsaved new document from exiting
For unsaved new documents, it was possible to exit the application
without saving the document.

If online either autosaved to the temporary location or the user
triggered a save, then cancelled not picking a final save location.

Well the two stage save mechanism is downright ugly and should be replaced!
Meanwhile this at least prevents unsaved new documents from exiting for
now.

Signed-off-by: Sarper Akdemir <sarper.akdemir@collabora.com>
Change-Id: I7ebe02c61194eb22fbd78aca0b7015af1971e3d5