Pass also the configured vendor to the CODA-W project
Will be used in a follow-up commit to construct a so-called
AppUserModelId for the case when running the app locally as a
developer. (When installed from the Microsoft Store the app will have
a such automatically.) An AppUserModelId is apparently needed for
functionality like getting files opened by the app into the Recent
Documents list.

Signed-off-by: Tor Lillqvist <tml@collabora.com>
Change-Id: I1e5aae3dd9e7fd0434b717f8e4846d886c174cf3