Introduce window.postMobileCall() for use when a value should be returned
On other platforms, at least Qt and macOS, the existing
postMobileMessage() can apparently be used for this already. See the
recently introduced code in
browser/src/control/Control.BackstageView.ts that does:

const result = await window.postMobileMessage('GETRECENTDOCS');

and that works at least on Qt. I am also told that it would work on
macOS, if only the CODA-M code would handle GETRECENTDOCS.

But in Windows, we need to set up the plumbing with a
Promise. Introduce a new function, window.postMobileCall() for these
cases where a return value is expected. On other platforms it is
identical to postMobileMessage(). When the CODA-W native code wants to
return the value, it should make the JS call the function
window.replyFromNativeToCall(). That will be in a follow-up commit.

Signed-off-by: Tor Lillqvist <tml@collabora.com>
Change-Id: Ic71b9f5648662be09e587804514d148b15608c09